Creacion de controladores y rutas para las tablas de actividad, alumno y maestro, ajuste de app.js
